+++ order-service/src/main/java/com/dili/orders/service/impl/WeighingStatementServiceImpl.java
@@ -21,26 +21,26 @@ import com.github.pagehelper.PageHelper;
 @Service
 public class WeighingStatementServiceImpl extends BaseServiceImpl<WeighingStatement, Long> implements WeighingStatementService {
 
-	public WeighingStatementMapper getActualDao() {
-		return (WeighingStatementMapper) getDao();
-	}
+    public WeighingStatementMapper getActualDao() {
+        return (WeighingStatementMapper) getDao();
+    }
 
-	@Override
-	public PageOutput<List<WeighingStatementAppletDto>> listApplet(WeighingStatementAppletQuery query) {
-		Integer page = query.getPage();
-		page = (page == null) ? Integer.valueOf(1) : page;
-		if (query.getRows() != null && query.getRows() >= 1) {
-			// 为了线程安全,请勿改动下面两行代码的顺序
-			PageHelper.startPage(page, query.getRows());
-		}
-		List<WeighingStatementAppletDto> list = this.getActualDao().listApplet(query);
-		Page<WeighingStatementAppletDto> pageList = (Page<WeighingStatementAppletDto>) list;
-		long total = pageList.getTotal();
-		return PageOutput.success().setData(list).setTotal((int) total).setPageNum(pageList.getPageNum()).setPageSize(pageList.getPageSize());
-	}
+    @Override
+    public PageOutput<List<WeighingStatementAppletDto>> listApplet(WeighingStatementAppletQuery query) {
+        Integer page = query.getPage();
+        page = (page == null) ? Integer.valueOf(1) : page;
+        if (query.getRows() != null && query.getRows() >= 1) {
+            // 为了线程安全,请勿改动下面两行代码的顺序
+            PageHelper.startPage(page, query.getRows());
+        }
+        List<WeighingStatementAppletDto> list = this.getActualDao().listApplet(query);
+        Page<WeighingStatementAppletDto> pageList = (Page<WeighingStatementAppletDto>) list;
+        long total = pageList.getTotal();
+        return PageOutput.success().setData(list).setTotal(total).setPageNum(pageList.getPageNum()).setPageSize(pageList.getPageSize());
+    }
 
-	@Override
-	public List<WeighingStatementAppletStateCountDto> stateCountStatistics(WeighingStatementAppletQuery query) {
-		return this.getActualDao().selectStateCount(query);
-	}
+    @Override
+    public List<WeighingStatementAppletStateCountDto> stateCountStatistics(WeighingStatementAppletQuery query) {
+        return this.getActualDao().selectStateCount(query);
+    }
 }
